Pass mainline module info file path when executed via Gantry.
Change the gantry_mode flag to mainline_modules_info_path.
This flag now accepts path to the json file instead of a boolean value.
If this flag is set, it should indicate that the script is executed via
Gantry.

Reason -
This script is copied to google3 using copybara and uses x20
fetch_artifact, which does not work in boq. Rapid which will
execute this script in google3 runs on boq.

Hence, the artifacts is fetched in google3 using android build
APIs. And the path of the download file is passed as parameter
to finalize_sdk script.

Bug: 321970705
Test: packages/modules/common/tools/finalize_sdk.py -f 14 -b 256848771 -c mainline -r 'Finalization Test' -m com.google.android.adservices 11609273
Test: packages/modules/common/tools/finalize_sdk.py -f 14 -b 256848771 -c mainline -p '/tmp/tmpjgh1delw/mainline-modules-info.json' -r 'Finalization Test' -m com.google.android.adservices 11609273

Change-Id: I19aef08c09ff1bc6433a1517c32b0fb64c304040