Revert "Revert "Update allow_deps.txt for refactored CaptivePort..."
Reason for revert: found the root cause of cts failure, and fix it with a new change which should be on top of this topic, so I revert revert this CL first and apply the fix on top of this.

Reverted changes: /q/submissionid:3206651-revert-3196237-CaptivePortalLoginLib-VOXHIDWDUB

Restore the previous CL which updates allow_deps.txt for refactored CaptivePortalLoginLib.

Apex-Size-Increase: no size change, the source files in the lib
such as CaptivePortalLoginActivity.java are already in the module.
Previous-Platform-Support: Yes
Aosp-First: Yes
Test-Info: Part of CaptivePortalLoginTests

Bug: 329361366
Test: TH
Change-Id: I909bfa2a051bdcb7d354bc2e2e08953bf151399b